pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "qrImageIndexerGUI"
version = "0.0.1"
description = "GUI interface for tool for indexing images with QR codes"
readme = "README.md"
authors = [{name = "Jonathan Pecar", email = "jonathan.pecar@gmail.com"}]
license = { file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["sorter", "qrcode", "photosorter"]
dependencies = [
        "qrImageIndexer>=0.7.1",
        "PyMuPDF",
        "pillow"
]
requires-python = ">=3.7"
[project.optional-dependencies]
dev = ["pytest", "bumpver", "setuptools", "pytest-mock"]

[project.urls]
Homepage = "https://github.com/jonpecar/qrCodeImageSorterGUI"
[tool.bumpver]
current_version = "0.0.1"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
'pyproject.toml' = [
    'current_version = "{version}"',
    'version = "{version}"',
]
'qrImageIndexerGUI/__init__.py' = [
    '__version__ = "{version}"',
]

